move services file to js directory
